@@ -3,22 +3,22 @@
   "description": "Gatsby default starter",
   "version": "1.0.0",
   "dependencies": {
-    "@types/styled-components": "^4.1.12",
-    "babel-preset-gatsby": "^0.1.4",
-    "gatsby": "^2.1.34",
-    "gatsby-image": "^2.0.15",
-    "gatsby-plugin-manifest": "^2.0.5",
-    "gatsby-plugin-offline": "^2.0.11",
-    "gatsby-plugin-react-helmet": "^3.0.0",
-    "gatsby-plugin-sharp": "^2.0.28",
-    "gatsby-plugin-typescript": "^2.0.11",
-    "gatsby-source-filesystem": "^2.0.26",
-    "gatsby-transformer-sharp": "^2.1.4",
-    "react": "^16.5.1",
-    "react-dom": "^16.5.1",
-    "react-helmet": "^5.2.0",
+    "@types/styled-components": "^4.1.15",
+    "babel-preset-gatsby": "^0.1.11",
+    "gatsby": "^2.7.1",
+    "gatsby-image": "^2.1.0",
+    "gatsby-plugin-manifest": "^2.1.1",
+    "gatsby-plugin-offline": "^2.1.1",
+    "gatsby-plugin-react-helmet": "^3.0.12",
+    "gatsby-plugin-sharp": "^2.1.0",
+    "gatsby-plugin-typescript": "^2.0.14",
+    "gatsby-source-filesystem": "^2.0.37",
+    "gatsby-transformer-sharp": "^2.1.19",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
+    "react-helmet": "^5.2.1",
     "react-loading": "^2.0.3",
-    "styled-components": "^4.1.1"
+    "styled-components": "^4.2.0"
   },
   "keywords": [
     "gatsby"
@@ -34,19 +34,19 @@
     "type-check:watch": "npm run type-check -- --watch"
   },
   "devDependencies": {
-    "prettier": "^1.14.2",
-    "@babel/cli": "^7.2.0",
-    "@babel/core": "^7.2.0",
-    "@babel/plugin-proposal-class-properties": "^7.2.1",
-    "@babel/plugin-proposal-object-rest-spread": "^7.2.0",
-    "@babel/preset-env": "^7.2.0",
+    "@babel/cli": "^7.4.4",
+    "@babel/core": "^7.4.5",
+    "@babel/plugin-proposal-class-properties": "^7.4.4",
+    "@babel/plugin-proposal-object-rest-spread": "^7.4.4",
+    "@babel/preset-env": "^7.4.5",
     "@babel/preset-typescript": "^7.1.0",
-    "tslint": "~5.11.0",
-    "tslint-config-prettier": "~1.15.0",
-    "typescript": "^3.2.2",
     "@types/react": "~16.4.18",
     "@types/react-dom": "~16.0.9",
-    "@types/react-helmet": "~5.0.7"
+    "@types/react-helmet": "~5.0.7",
+    "prettier": "^1.17.1",
+    "tslint": "~5.11.0",
+    "tslint-config-prettier": "~1.15.0",
+    "typescript": "^3.4.5"
   },
   "repository": {
     "type": "git",
