@@ -1,26 +1,9 @@
 import React from 'react'
-import styled from 'styled-components'
 
-const LinkList = styled.ul`
-  display: inline-flex;
-  padding-left: 0;
-  padding-bottom: 16px;
-  justify-content: space-between;
-`
+import { Footer, LinkItem, LinkList, SmallLink } from './styles'
 
-const LinkItem = styled.a`
-  padding-right: 10px;
-  font-size: 16px;
-  font-family: 'Montserrat', sans-serif;
-`
-
-const SmallLink = styled.a`
-  text-transform: uppercase;
-  font-family: 'Montserrat', sans-serif;
-`
-
-const LayoutFooter = () => (
-  <footer style={{ fontWeight: '300' }}>
+const LayoutFooter = (): JSX.Element => (
+  <Footer>
     -
     <div>
       <LinkList>
@@ -36,7 +19,7 @@ const LayoutFooter = () => (
         Github Repo
       </SmallLink>
     </small>
-  </footer>
+  </Footer>
 )
 
 export default LayoutFooter
